fix:  - cache object is not shared between instances
