+++ package.json
@@ -24,7 +24,7 @@
         "test-coverage": "jest --coverage"
     },
     "peerDependencies": {
-        "axios": "~0.18.0 || ~0.19.1 || ~0.20.0 || ~0.21.1"
+        "axios": ">= 0.18 < 0.19.0 || >= 0.19.1"
     },
     "devDependencies": {
         "@types/jest": "^24.9.1",
